+++ girderformindlogger/models/protocol.py
@@ -33,7 +33,6 @@ from girderformindlogger.models.folder import Folder as FolderModel
 from girderformindlogger.models.user import User as UserModel
 from girderformindlogger.utility.progress import noProgress, setResponseTimeLimit
 
-
 class Protocol(FolderModel):
     def importUrl(self, url, user=None, refreshCache=False):
         """
@@ -116,3 +115,8 @@ class Protocol(FolderModel):
                 raise ValidationException(
                     "Invalid Protocol ID."
                 )
+
+    def createProtocol(self, document, user):
+        from girderformindlogger.utility import jsonld_expander
+
+        return jsonld_expander.loadFromSingleFile(document, user)